Move replace string dialog to new package (and remove duplicate JavaFX conversion)
